requirements.txt
Django~=5.1.1
psycopg2~=2.9.9
psycopg2-binary~=2.9.9
djangorestframework~=3.15.2